+++ apps/mobile/index.js
@@ -2,5 +2,8 @@ import './polyfills';
 import registerRootComponent from 'expo/src/launch/registerRootComponent';
 
 import App from './App';
+import * as Sentry from '@sentry/react-native';
+import { initSentry } from './utils/sentry';
 
-registerRootComponent(App);
+initSentry();
+registerRootComponent(Sentry.wrap(App));